chore: update autospy options
