+++ tests/spec/all.autospy-options.spec.ts
@@ -1,41 +1,30 @@
-import { Tree } from '@angular-devkit/schematics';
-import { SchematicTestRunner } from '@angular-devkit/schematics/testing';
-import { collectionPath } from './common';
+import { setupBase } from './common';
 
 describe('autoSpy options', () => {
-    let tree = Tree.empty();
-    const fileName = 'test.ts';
-    const testFileName = 'test.spec.ts'
-
-    beforeEach(() => {
-        tree = Tree.empty();
-    })
+    const file = 'test.ts';
+    const folder = 'all.autospy-options';
 
     it('when creating a new spec it should use the autospy options from command line overwriting the option file', async () => {
         // arrange
-        const runner = new SchematicTestRunner('schematics', collectionPath);
-        tree.create(fileName, fileContents());
+        const { run, fullFileName, add, testFileName, getFilePath } = setupBase(folder, file);
+        add(fullFileName);
         // act
-        const r = await runner.runSchematicAsync(
-            'spec',
-            { name: fileName, config: 'tests/spec/test-data/config-autospy.json', autoSpyLocation:'@testing/autospy' },
-            tree
-        ).toPromise();
+        const result = await run({
+            name: fullFileName,
+            config: getFilePath('config-autospy.json'),
+            autoSpyLocation: '@testing/autospy',
+        });
         // assert
-        const res = r.read(testFileName)?.toString();
+        const res = result.read(testFileName)?.toString();
         expect(res).toMatch(`import { autoSpy } from '@testing/autospy';`);
     });
 
     it('when creating a new spec it should use the autospy options from option file', async () => {
         // arrange
-        const runner = new SchematicTestRunner('schematics', collectionPath);
-        tree.create(fileName, fileContents());
-        // act
-        const r = await runner.runSchematicAsync(
-            'spec',
-            { name: fileName, config: 'tests/spec/test-data/config-autospy.json' },
-            tree
-        ).toPromise();
+        const { run, fullFileName, add, testFileName, getFilePath } = setupBase(folder, file);
+        add(fullFileName);
+        //act
+        const r = await run({ name: fullFileName, config: getFilePath('config-autospy.json') });
         // assert
         const res = r.read(testFileName)?.toString();
         expect(res).toMatch(`import { autoSpy } from '@some/path';`);
@@ -43,15 +32,17 @@ describe('autoSpy options', () => {
 
     it('when updating a spec it should use the autospy options from command line overwriting the option file', async () => {
         // arrange
-        const runner = new SchematicTestRunner('schematics', collectionPath);
-        tree.create(fileName, fileContents());
-        tree.create(testFileName, testFileContents());
-        // act
-        const r = await runner.runSchematicAsync(
-            'spec',
-            { name: fileName, update: true, config: 'tests/spec/test-data/config-autospy.json', autoSpyLocation:'@testing/autospy' },
-            tree
-        ).toPromise()
+        // arrange
+        const { run, fullFileName, add, testFileName, getFilePath } = setupBase(folder, file);
+        add(fullFileName);
+        add(testFileName);
+        //act
+        const r = await run({
+            name: fullFileName,
+            update: true,
+            config: getFilePath('config-autospy.json'),
+            autoSpyLocation: '@testing/autospy',
+        });
         // assert
         const res = r.read(testFileName)?.toString();
         expect(res).toMatch(`import { autoSpy } from '@testing/autospy';`);
@@ -59,60 +50,17 @@ describe('autoSpy options', () => {
 
     it('when updating a spec it should use the autospy options from the option file', async () => {
         // arrange
-        const runner = new SchematicTestRunner('schematics', collectionPath);
-        tree.create(fileName, fileContents());
-        tree.create(testFileName, testFileContents());
-        // act
-        const r = await runner.runSchematicAsync(
-            'spec',
-            { name: fileName, update: true, config: 'tests/spec/test-data/config-autospy.json' },
-            tree
-        ).toPromise()
+        const { run, fullFileName, add, testFileName, getFilePath } = setupBase(folder, file);
+        add(fullFileName);
+        add(testFileName);
+        //act
+        const r = await run({
+            name: fullFileName,
+            config: getFilePath('config-autospy.json'),
+            update: true,
+        });
         // assert
         const res = r.read(testFileName)?.toString();
         expect(res).toMatch(`import { autoSpy } from '@some/path';`);
     });
-
 });
-
-
-function fileContents() {
-    return `import {Dep} from './dep';
-    class MyClass {
-        constructor(v: Dep) {}
-        method() {
-            return 'result'
-        }
-    }`
-}
-
-function testFileContents() {
-    return `';
-    import { MyClass } from './test';
-
-    describe('MyClass', () => {
-      it('when method is called it should', () => {
-        // arrange
-        const { build } = setup().default();
-        const m = build();
-        // act
-        m.method();
-        // assert
-        // expect(m).toEqual
-      });
-
-    });
-
-    function setup() {
-      const builder = {
-        default() {
-          return builder;
-        },
-        build() {
-          return new MyClass();
-        }
-      };
-
-      return builder;
-    }`
-}
