+++ tests/spec/test-data/all.autospy-options/config-autospy.json
