+++ tests/spec/test-data/all.autospy-options/test.ts
@@ -1,0 +1,7 @@
+import { Dep } from './dep';
+class MyClass {
+    constructor(v: Dep) {}
+    method() {
+        return 'result';
+    }
+}
